Revert "Split Primary and Replica DB by using trait"
This reverts commit c94c3cca356b33108c929d04476f4d93db4422d0.
