+++ crates/full-node/sov-sequencer/src/preferred/mod.rs
@@ -27,7 +27,9 @@ use crate::preferred::rpc_errors::{cant_fit_tx, rate_limit, replica_mode, shut_d
 use crate::preferred::timestamp::{update_timestamp_task, TimingOracleConfigWithPrivateKey};
 use async_trait::async_trait;
 use batch_size_tracker::BatchSizeTracker;
-use db::{ReadBatch, ReadBlob};
+use db::postgres::PostgresBackend;
+use db::rocksdb::RocksDbBackend;
+use db::{PreferredSequencerDb, ReadBatch, ReadBlob};
 use derive_more::Deref;
 use futures::Stream;
 pub use initialization::Builder;
